+++ src/bootstrapRoot.tsx
@@ -3,13 +3,15 @@
 'use client';
 
 import type { PropsWithChildren } from 'react';
-import React from 'react';
+import React, { useContext } from 'react';
 import type { RecoilState } from 'recoil';
 import {
   useGotoRecoilSnapshot,
   useRecoilSnapshot,
   useRecoilStateLoadable,
 } from 'recoil';
+
+import { BootstrapRootsInScopeContext } from './util';
 
 interface LocalizedStateProps<BootstrapData> {
   bootstrapData: BootstrapData;
@@ -26,6 +28,12 @@ export function BootstrapRoot<BootstrapData>({
   // been initialized with bootstrap data yet, and that we should do so now.
   const [bootstrapDataLoadable] = useRecoilStateLoadable(rootAtom);
 
+  // We get the list of known root atoms above this one. If any exist, then that
+  // means this root is nested inside of another root. We need to pass along
+  // the list of all roots to the nested context so that value hooks can check
+  // if they are being called in the right component tree
+  const parentBootstrapRoots = useContext(BootstrapRootsInScopeContext);
+
   // We use snapshots to set the root atom value instead of the typical
   // `useRecoilState` setter because the latter is asynchronous and would result
   // in bootstrap data not being available for the first render.
@@ -38,7 +46,18 @@ export function BootstrapRoot<BootstrapData>({
       }),
     );
   }
+
   // TODO: add some logic that puts the root atom back in a loadable state on
   // component unmount here.
-  return <>{children}</>;
+
+  return (
+    <BootstrapRootsInScopeContext.Provider
+      // We add this bootstrap root's root atom to the list of known roots.
+      // Turns out, we can nest the same provider multiple times, and it scopes
+      // itself so that we always get the right value in the right place.
+      value={[...parentBootstrapRoots, rootAtom]}
+    >
+      {children}
+    </BootstrapRootsInScopeContext.Provider>
+  );
 }