+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## ?
+
+- Fixed bug where bootstrapped atoms can be accessed outside of their bootstrap roots when still accessed inside of the Recoil root
+
 ## 1.0.1 (06/23/2023)
 
 - Fixed bugs where root atoms weren't rehydrating properly in some cases in Next.js 13 app-router